исправил getdata и sendData (+колбэки)
