+++ js/server.js
@@ -1,30 +1,36 @@
 import {createStandartMarker} from './map.js';
-import {getRandomPositiveInteger, showSuccessAlert, showErrorAlert} from './utils.js';
-import {resetForm} from './form-validate.js';
+import {getRandomPositiveInteger, showErrorAlert} from './utils.js';
 
 const SIMILAR_OFFERS_COUNT = getRandomPositiveInteger(1, 10);
-const submitButton = document.querySelector('.ad-form__submit');
+
 const filterFormElement = document.querySelector('.map__filters');
 
 const renderOffers = (offers) => {
   offers.slice(0, SIMILAR_OFFERS_COUNT).forEach((offer) => createStandartMarker(offer));
 };
 
-const getData = (onSuccess) => {
+const getData = (onSuccess, onError) => {
 
   filterFormElement.classList.add('ad-form--disabled');
 
   fetch('https://26.javascript.pages.academy/keksobooking/data')
-    .then((response) => response.json())
-    .then((offers) => onSuccess(offers))
+    .then((response) => {
+      if (response.ok) {
+        return response.json();
+      } else {
+        onError();
+      }
+    })
+    .then((offers) => {
+      onSuccess(offers);
+    })
     .catch(() => showErrorAlert('Не удалось загрузить данные'))
     .finally(() => filterFormElement.classList.remove('ad-form--disabled'));
 };
 
-getData(renderOffers);
+getData(renderOffers, () => showErrorAlert('Не удалось загрузить данные'));
 
-const sendData = (body) => {
-  submitButton.disabled = true;
+const sendData = (body, onSuccess, onError) => {
   fetch(
     'https://26.javascript.pages.academy/keksobooking',
     {
@@ -34,14 +40,12 @@ const sendData = (body) => {
   )
     .then((response) => {
       if (response.ok) {
-        showSuccessAlert();
-        resetForm();
+        onSuccess();
       } else {
-        showErrorAlert('Не удалось отправить данные');
+        onError();
       }
     })
-    .catch(showErrorAlert)
-    .finally(submitButton.disabled = false);
+    .catch(showErrorAlert);
 };
 
 export {sendData};
